dsdsd

+++ Tables/DimFixFinPostingType.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[DimFixFinPostingType]
 (
-	TransTypeKey tinyint NOT NULL  PRIMARY KEY,
+	TransTypeKey  int NOT NULL  PRIMARY KEY,
 	TransTypeName nvarchar(50) Not null
 )
